+++ fstat64.c
@@ -68,7 +68,7 @@ exit:
 
 int __fstat64 (int __fd, struct stat64 *__buf) __THROW __nonnull ((2));
 weak_alias(fstat64, __fstat64);
-int __fstat_time64 (int __fd, struct stat64 *__buf) __THROW __nonnull ((2));
-weak_alias(fstat64, __fstat_time64);
+int __fstat64_time64 (int __fd, struct stat64 *__buf) __THROW __nonnull ((2));
+weak_alias(fstat64, __fstat64_time64);
 #endif
 #endif