+++ fstatat64.c
@@ -78,9 +78,9 @@ int __fstatat64 (int __fd, const char *__restrict __file,
 		 struct stat64 *__restrict __buf, int __flag)
      __THROW __nonnull ((2, 3));
 weak_alias(fstatat64, __fstatat64);
-int __fstatat_time64 (int __fd, const char *__restrict __file,
-		      struct stat64 *__restrict __buf, int __flag)
+int __fstatat64_time64 (int __fd, const char *__restrict __file,
+			struct stat64 *__restrict __buf, int __flag)
      __THROW __nonnull ((2, 3));
-weak_alias(fstatat64, __fstatat_time64);
+weak_alias(fstatat64, __fstatat64_time64);
 #endif
 #endif