+++ stat64.c
@@ -25,8 +25,9 @@ int stat64(const char *path, struct stat64 *statbuf)
 int __stat64 (const char *__restrict __file,
 	      struct stat64 *__restrict __buf) __THROW __nonnull ((1, 2));
 weak_alias(stat64, __stat64);
-int __stat_time64 (const char *__restrict __file,
-		   struct stat64 *__restrict __buf) __THROW __nonnull ((1, 2));
-weak_alias(stat64, __stat_time64);
+int __stat64_time64 (const char *__restrict __file,
+		     struct stat64 *__restrict __buf)
+     __THROW __nonnull ((1, 2));
+weak_alias(stat64, __stat64_time64);
 #endif
 #endif